为ISpider创建抽象类。
